requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.30
alembic==1.13.2
mysqlclient==2.2.0
PyMySQL==1.1.0
pydantic==2.5.0
email-validator==2.1.0
python-jose[cryptography]
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-dotenv==1.0.0
mypy==1.7.1
black==23.11.0
isort==5.12.0
structlog==23.2.0
pydantic_settings
requests-oauthlib==1.3.1